move utils to composer/samplecode
