+++ src/test/java/com/google/api/generator/gapic/composer/samplecode/BUILD.bazel
@@ -3,22 +3,22 @@ load("@rules_java//java:defs.bzl", "java_test")
 package(default_visibility = ["//visibility:public"])
 
 TESTS = [
-    "JavaStyleTest",
+    "SampleCodeJavaFormatterTest",
 ]
 
 filegroup(
-    name = "utils_files",
+    name = "samplecode_files",
     srcs = ["{0}.java".format(f) for f in TESTS],
 )
 
 [java_test(
     name = test_name,
     srcs = ["{0}.java".format(test_name)],
-    test_class = "com.google.api.generator.gapic.utils.{0}".format(test_name),
+    test_class = "com.google.api.generator.gapic.composer.samplecode.{0}".format(test_name),
     deps = [
         "//src/main/java/com/google/api/generator/engine/ast",
         "//src/main/java/com/google/api/generator/engine/writer",
-        "//src/main/java/com/google/api/generator/gapic/utils",
+        "//src/main/java/com/google/api/generator/gapic/composer/samplecode",
         "@com_google_truth_truth//jar",
         "@junit_junit//jar",
     ],
